decomiss-datanodes-hadoop/requirements.txt
argparse==1.2.1
python2-pythondialog==3.4.0

